dealing with merge conflicts from #92
