+++ .gitignore
@@ -1,3 +1,4 @@
 dist/
 hauser
 *.toml
+.idea
