+++ go.mod
@@ -6,11 +6,10 @@ require (
 	cloud.google.com/go/pubsub v1.0.1 // indirect
 	github.com/BurntSushi/toml v0.3.1
 	github.com/aws/aws-sdk-go v1.23.18
-	github.com/gordonklaus/ineffassign v0.0.0-20200809085317-e36bfde3bb78 // indirect
 	github.com/lib/pq v1.2.0
-	github.com/nishanths/predeclared v0.0.0-20200524104333-86fad755b4d3 // indirect
 	github.com/pkg/errors v0.8.1
 	github.com/stretchr/testify v1.4.0 // indirect
+	golang.org/x/tools v0.0.0-20200522201501-cb1345f3a375 // indirect
 	honnef.co/go/tools v0.0.1-2020.1.5 // indirect
 )
